Fix bounding issue in remove
Previously, we only passed the in-bound (actual offset - piece start) to
successive calls to remove when a deletion spans multiple pieces. This
is incorrect. We need to account for the edits already handled in the
current piece and get whatever the difference is between those and the
extent of the piece (edit length - in bounds) (since we only delete
leftwards).