Add move and substitute operations; update insert
The two new operations here are convenience functions for moving pieces
of the original buffer in a table or for substituting some range in the
buffer with another string. One could use this, for instance, to support
things like find/replace.

This commit also updates the signature of insert to take a string
instead of a ref to a string.